requirements.txt
asgiref==3.4.1
Deprecated==1.2.13
django-filter==21.1
django-jazzmin==2.4.8
Django==3.2.9
djangorestframework==3.12.4
flake8==4.0.1
importlib-metadata==4.8.2
Markdown==3.3.6
mccabe==0.6.1
Pillow==8.4.0
pip==21.3.1
psycopg2==2.9.2
psycopg2-binary==2.9.2
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2021.3
redis==4.0.1
requests==2.26.0
setuptools==59.2.0
sqlparse==0.4.2
wrapt==1.13.3
zipp==3.6.0
